chore(src): resolve all inspection problems from problems/
Run clang-tidy -fix across src/, add targeted NOLINTs for non-fixable

checks (recursion, exception-escape, redundant declarations), and remove

safe unused includes from .cpp files. Header umbrella includes are kept

intentionally. Fix follow-up build issues: add missing

transparentDebugGraphicsPipeline field, align const qualifiers, make

RtxShadowPipeline group-index helpers static, default destructors.

Build green; 43/43 ctest pass; zero clang-tidy warnings in src/.

+++ src/render/RendererDrawFrame.cpp
@@ -372,7 +372,7 @@ SDL_AppResult DrawFrame(
 				0,
 				0,
 				0,
-				static_cast<int32_t>(voxelWorld->chunkSize),
+				voxelWorld->chunkSize,
 			};
 			worldGenPush.chunkCountAndFlags = {
 				worldGenChunkCount,
